// src/model/interfaces/VerificationErrorCode.ts
export enum VerificationErrorCode {
    VERIFIED_ADDRESSES_BUT_NO_THUMBPRINT = 401_001,
    VERIFIED_ADDRESSES_KEYS_DO_NOT_MATCH_THUMBPRINT = 401_002,
    VERIFIED_ADDRESSES_AND_ADDRESSES_DIFFER_IN_LENGTH = 401_003,
    VERIFIED_ADDRESSES_AND_ADDRESSES_DIFFER_CONTENT = 401_004,
    SYSTEM_ERROR_VERIFYING = 500_001,
}